// site/src/app/main/pages/projects/pages/project-view/models/project-page-view.const.ts
export enum PROJECT_PAGE_VIEW {
  ABOUT = 'about',
  PHASES = 'phases',
  DISCIPLINES = 'disciplines',
  DISCUSSIONS = 'discussions',
  TIMELINE = 'timeline',
  RESOURCES = 'resources',
  UPLOAD = 'upload',
  SETTINGS = 'settings',
}

export type PROJECT_PAGE_VIEW_TYPE =
  | PROJECT_PAGE_VIEW.ABOUT
  | PROJECT_PAGE_VIEW.PHASES
  | PROJECT_PAGE_VIEW.DISCIPLINES
  | PROJECT_PAGE_VIEW.DISCUSSIONS
  | PROJECT_PAGE_VIEW.RESOURCES
  | PROJECT_PAGE_VIEW.TIMELINE
  | PROJECT_PAGE_VIEW.SETTINGS;